requirements.txt
requests
requests-html
BeautifulSoup4
google
python-nmap
fpdf2
python-whois